% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asrbatch.R
\name{asrbatch}
\alias{asrbatch}
\title{asreml batch analysis}
\usage{
asrbatch(data,factorN,traitN,
                 FMod=NULL,RMod=NULL, EMod=NULL,
                 mulT=NULL, mulN=NULL,mulR=NULL,
                 corM=NULL,corMout=FALSE,
                 order=FALSE,plot.matrix=FALSE,all.result=FALSE,
                 rgFormula=NULL,pformula=NULL,pformula1=NULL,
                 pformula2=NULL,pformula3=NULL,pformula4=NULL,
                 sigFormula=NULL,maxit=NULL,asrV=3,
                 SPmodel=FALSE,digit=4,family=NULL,
                 ped=NULL,pedinv=NULL,ginverse=NULL)
}
\arguments{
\item{data}{aim dataset}

\item{factorN}{A vector with sites of all factors.}

\item{traitN}{A vector with sites of all traits.}

\item{FMod}{Fixed mode.}

\item{RMod}{Randomed variance structure (G structure).}

\item{EMod}{Error variance structure (R) for multi-trait model.}

\item{mulT}{Value "T" or "TRUE" for multi-trait model, "F"(default).}

\item{mulN}{Number of trait for one model in multi-trait analysis,2(default).}

\item{mulR}{Logical Value "T" or "TRUE" to count corr/error matrix, only works for bitrait, "F"(default).}

\item{corM}{Logical Value "T" or "TRUE" for corr model, "F"(default).}

\item{corMout}{Logical Value "T" or "TRUE" to output corr matrix, "F"(default).}

\item{order}{Logical Value "T" or "TRUE" to reorder corr matrix, "T"(default).}

\item{plot.matrix}{Logical Value "T" or "TRUE" with corMout (also 'T')to output corr matrix for plotting, "F"(default).}

\item{all.result}{Logical Value "T" or "TRUE" to return each asreml run's all results,"F"(default).}

\item{rgFormula}{formula for corr, etc.}

\item{sigFormula}{Logical Value "T" or "TRUE" to test signif for formula, "F"(default).}

\item{maxit}{Maximum number of iterations.}

\item{asrV}{asreml version 3(default V3), or 4 (new version V4).}

\item{SPmodel}{for spatial analysis model.}

\item{digit}{index to control decimal number.}

\item{family}{A list of functions and expressions for defining the link and variance functions.}

\item{ped}{Value "T" or "TRUE" for animal model with pedigree, "F"(default).}

\item{pedinv}{A G-inverse matrix for pedigree from ASReml.Ainverse()--V3 or ainverse()--V4.}

\item{ginverse}{A named list with each component identifying a G-inverse matrix, working for V3.}

\item{formula}{formula for h2 or corr.}

\item{formula1}{formula for h2 or corr.}

\item{formula2}{formula for h2 or corr.}

\item{formula3}{formula for h2 or corr.}

\item{formula4}{formula for h2 or corr.}
}
\value{
the result is returned directly.
}
\description{
\code{asrbatch} This function carries out batch analysis for  
mult-trait with same model and also output heritability etc. in asreml package.
}
\details{
asreml batch analysis
}
\examples{
\dontrun{
library(AAFMM)

##### example I for dataset without pedigree in ASReml-R V3.0
library(asreml)

data(PrSpa,package='AAFMM')
df<-PrSpa

# exmaple 1.1 single trait model
df1<-subset(df,Spacing==3)

asrbatch(data=df1,factorN=1:5,traitN=c(9:13),
          FMod=y~1+Rep+Plot,
          RMod=~Fam, 
          #family=asreml.gaussian(),
          pformula=h2 ~ 4 * V1/(V1+V2))
          
# exmaple 1.2 us model 
asrbatch(data=df1,factorN=1:5,traitN=c(10:13),
          FMod=cbind(y1,y2)~trait+trait:Rep,
          RMod=~us(trait):Fam,
          EMod=~units:us(trait),
          #family=c(asreml.gaussian(),asreml.gaussian()),
          mulT=TRUE,mulN=2,mulR=TRUE,corMout=TRUE,
          pformula=gcorr ~ V2/sqrt(V1*V3),
          pformula1=h2.A ~ 4*V1/(V1+V5),
          pformula2=h2.B ~ 4*V3/sqrt(V3+V7)) 
          

# exmaple 1.3 corr model
asrbatch(data=df1,factorN=1:5,traitN=c(10:13),
             FMod=cbind(y1,y2,y3)~trait+trait:Rep,
             RMod=~corgh(trait):Fam,
             EMod=~units:us(trait), 
             maxit=30,
             mulT=TRUE,mulN=3,mulR=TRUE,corM=TRUE)  

       
##### example II for dataset with pedigree in ASReml-R V3.0
data(dfm2,package='AAFMM')

ped<-dfm2[,1:3]  
ainv<-asreml.Ainverse(ped)$ginv

dfm2a<-subset(dfm2,Spacing==3)  

# example 2.1 single trait model
asrbatch(data=dfm2a,factorN=1:6,traitN=c(7:14),
          FMod=y~1+Rep,
          RMod=~ped(TreeID),
          ped=TRUE,pedinv=ainv, 
          ginverse=list(TreeID=ainv),
          pformula=h2 ~ V1/(V1+V2)) 
          
# exmaple 2.2 us model 
asrbatch(data=dfm2a,factorN=1:6,traitN=c(10:14),
             FMod=cbind(y1,y2)~trait+trait:Rep,
             RMod=~us(trait):ped(TreeID),
             EMod=~units:us(trait),maxit=40,
             mulT=TRUE,mulN=2,mulR=TRUE,corMout=FALSE,
             ped=TRUE,pedinv=ainv,
             ginverse=list(TreeID=ainv),
             pformula=gcorr ~ V2/sqrt(V1*V3),
             pformula1=h2.A ~ V1/(V1+V5),
             pformula2=h2.B ~ V3/(V3+V7)) 
   
# exmaple 2.3  corr model
asrbatch(data=dfm2a,factorN=1:6,traitN=c(10:14),
             FMod=cbind(y1,y2)~trait+trait:Rep,
             RMod=~corgh(trait):ped(TreeID),
             EMod=~units:us(trait),maxit=40,
             mulT=TRUE,mulN=2,corM=TRUE,
             ped=TRUE,pedinv=ainv,
             ginverse=list(TreeID=ainv))
             
####### For asreml-R V4.0                       
library(asreml4)
library(AAFMM)

##### example III for dataset without pedigree in ASReml-R V4.0
data(PrSpa)
df<-PrSpa

# exmaple 3.1  single trait model
df1<-subset(df,Spacing==3)

asrbatch(data=df1,factorN=1:5,traitN=c(9:13),
          FMod=y~1+Rep+Plot,
          RMod=~Fam,
          asrV=4,
          #family=asr_gaussian(),
          pformula=h2 ~ 4 * V1/(V1+V2))
          
# exmaple 3.2 us model 
asrbatch(data=df1,factorN=1:5,traitN=c(10:13),
          FMod=cbind(y1,y2)~trait+trait:Rep,
          RMod=~us(trait):Fam,
          EMod=~units:us(trait),
          asrV=4,
          #family=list(asr_gaussian(),asr_gaussian()), 
          mulT=TRUE,mulN=2,mulR=TRUE,corMout=TRUE,
          pformula=gcorr ~ V2/sqrt(V1*V3),
          pformula1=h2.A ~ 4*V1/(V1+V5),
          pformula2=h2.B ~ 4*V3/sqrt(V3+V7)) 
          

# exmaple 3.3  corr model
asrbatch(data=df1,factorN=1:5,traitN=c(10:13),
             FMod=cbind(y1,y2,y3)~trait+trait:Rep,
             RMod=~corgh(trait):Fam,
             EMod=~units:us(trait),
             maxit=30,asrV=4,
             mulT=TRUE,mulN=3,mulR=TRUE,corM=TRUE) 


# exmaple 3.4 spatial model for single trait
data('sp',package='AAFMM') 
            
asrbatch(data=sp,factorN=1:8,traitN=c(9:10),
              FMod=y~1+Rep,
              RMod=~units+Mum,
              EMod=~ar1(Row):ar1(Col),
              maxit=30,SPmodel=TRUE,asrV=4,
              pformula=h2 ~ 4 * V1/(V1+V2)) 


# exmaple 3.5  spatial model for double traits
asrbatch(data=sp,factorN=1:8,traitN=c(9:10),
              FMod=cbind(y1,y2)~trait+trait:Rep,
              RMod=~diag(trait):units+us(trait):Mum,
              EMod=~ar1(Row):ar1(Col):diag(trait),
              maxit=30,SPmodel=TRUE,asrV=4,
              mulT=TRUE,mulN=2,
              mulR=FALSE,corMout=FALSE,
              pformula=h2.A ~ 4 * V1/(V1+V4))


##### example IV for dataset with pedigree in ASReml-R V4.0
data(dfm2,package='AAFMM')

ped<-dfm2[,1:3]  
ainv<-ainverse(ped)

dfm2a<-subset(dfm2,Spacing==3)  

# example 4.1 sigle trait model
asrbatch(data=dfm2a,factorN=1:6,traitN=c(7:14),
          FMod=y~1+Rep,
          RMod=~vm(TreeID,ainv),
          ped=TRUE,pedinv=ainv, asrV=4, 
          pformula=h2 ~ V1/(V1+V2)) 
          
# exmaple 4.2 us model 
asrbatch(data=dfm2a,factorN=1:6,traitN=c(10:14),
             FMod=cbind(y1,y2)~trait+trait:Rep,
             RMod=~us(trait):vm(TreeID,ainv),
             EMod=~units:us(trait),
             mulT=TRUE,mulN=2,
             mulR=TRUE,corMout=FALSE,
             ped=TRUE,pedinv=ainv,
             maxit=40,asrV=4,
             pformula=gcorr ~ V2/sqrt(V1*V3),
             pformula1=h2.A ~ V1/(V1+V5),
             pformula2=h2.B ~ V3/(V3+V7)) 
   
# exmaple 4.3  corr model
asrbatch(data=dfm2a,factorN=1:6,traitN=c(10:14),
             FMod=cbind(y1,y2)~trait+trait:Rep,
             RMod=~corgh(trait):vm(TreeID,ainv),
             EMod=~units:us(trait),maxit=40,
             mulT=TRUE,mulN=2,corM=TRUE,
             ped=TRUE,pedinv=ainv,asrV=4)
             

# exmaple 4.4 for spatial model-- single trait
data('sp',package='AAFMM')

ped<-sp[,1:3]
ainv<-ainverse(ped) 
            
asrbatch(data=sp,factorN=1:8,traitN=c(9:10),
              FMod=y~1+Rep,
              RMod=~units+vm(Tree,ainv),
              EMod=~ar1(Row):ar1(Col),
              maxit=30,SPmodel=TRUE,
              ped=TRUE,pedinv=ainv,asrV=4,
              pformula=h2 ~ V2/(V1+V2)) 


# exmaple 4.5 for spatial model-- double traits
asrbatch(data=sp,factorN=1:8,traitN=c(9:10),
              FMod=cbind(y1,y2)~trait+trait:Rep,
              RMod=~diag(trait):units+us(trait):vm(Tree,ainv),
              EMod=~ar1(Row):ar1(Col):diag(trait),
              maxit=30,SPmodel=TRUE,
              ped=TRUE,pedinv=ainv,asrV=4,
              mulT=TRUE,mulN=2,
              mulR=FALSE,corMout=FALSE,
              pformula=h2.A ~ V3/(V1+V3),
              pformula1=h2.B ~ V5/(V2+V5)) 
              
}

                            
}
\references{
Yuanzhen Lin. R & ASReml-R Statistics. China Forestry Publishing House. 2016  
AAFMM website:https://github.com/yzhlinscau/AAFMM
}
\author{
Yuanzhen Lin <yzhlinscau@163.com>
}
