% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dial.combn.R
\name{dial.combn}
\alias{dial.combn}
\title{Recipro and family code of diallel mating.}
\usage{
dial.combn(lbls, N)
}
\arguments{
\item{lbls}{An english character for genotype.}

\item{N}{diallel mating number.}
}
\value{
the result is returned directly.
}
\description{
\code{dial.combn} This function generates the recipro and family code from a diallel mating.
}
\details{
The input data is a character with a total number for diallel mate.
}
\examples{
 
library(AAFMM)
dial.combn("G",8)


}
\references{
Yuanzhen Lin. R & ASReml-R Statistics. China Forestry Publishing House. 2016 
AAFMM website:https://github.com/yzhlinscau/AAFMM
}
\author{
Yuanzhen Lin <yzhlinscau@163.com>
}
