% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap1.R
\name{heatmap1}
\alias{heatmap1}
\title{Plot correlation or matrix heatmap.}
\usage{
heatmap1(df,type='matrix', df.label=NULL,gtitle=NULL,
                   Nbreaks=NULL,Sig= FALSE,order=FALSE,
                   theme.1=NULL,data.only = FALSE)
}
\arguments{
\item{df}{A data.frame or matrix containing only numeric data.}

\item{type}{Identify df to be 'matrix'(default) or 'data'.}

\item{df.label}{A matrix for heatmap labels.}

\item{gtitle}{legend title.}

\item{Nbreaks}{A number controls legend breaks.}

\item{Sig}{Logical, if TRUE put pvalue and sig level for heatmap labels.}

\item{order}{Logical, if TRUE reorder correlation matrix.}

\item{theme.1}{A theme setting for ggplot2.}

\item{data.only}{Logical, if TRUE returns correlation and pvalue.}
}
\description{
\code{heatmap1} This function Function creates a correlation heatmap 
                 using ggplot2 given a data.frame or matrix.
}
\examples{

library(AAFMM)

mydata<-mtcars

## plot data.frame
heatmap1(mydata,type='data',Sig=FALSE,Nbreaks = 8)

heatmap1(mydata,type='data',Sig=TRUE,Nbreaks = 8)

## plot correlation matrix
mcor <- round(cor(mydata),2)
cormat <- melt(mcor)

heatmap1(cormat,Sig=FALSE,Nbreaks = 8)
heatmap1(cormat,Sig=TRUE,Nbreaks = 8)

## plot any squared matrix
set.seed(2018)
label<-matrix(runif(121),nrow=11)
label<-round(label,2)

heatmap1(label)

  
}
\references{
Yuanzhen Lin. R & ASReml-R Statistics. China Forestry Publishing House. 2016 
AAFMM website:https://github.com/yzhlinscau/AAFMM
}
\author{
Yuanzhen Lin <yzhlinscau@163.com>
}
