% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ir.sp.R
\docType{data}
\name{ir.sp}
\alias{ir.sp}
\title{Irregular spatial dataset}
\format{one dataset with 16 variables
\describe{
 \item{Provence}{3 levels}
 \item{Plustree}{Plustree code}
 \item{Family}{Family code2}
 \item{Block}{2 levels}
 \item{Row}{15 levels}
 \item{Col}{30 levels}
 \item{h05}{a numeric vector}
 \item{cw05}{a numeric vector}
 \item{gd05}{a numeric vector}
 \item{d05}{a numeric vector}
 \item{h07}{a numeric vector}
 \item{cw07}{a numeric vector}
 \item{d07}{a numeric vector}
 \item{h10}{a numeric vector}
 \item{d10}{a numeric vector}
 \item{v10}{a numeric vector}
}}
\usage{
data(ir.sp)
}
\description{
An irregular spatial dataset example to be transformed to regular spatial data.
}
\keyword{datasets}
