% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ir2r.sp.R
\name{ir2r.sp}
\alias{ir2r.sp}
\title{Irregular data to regular spatial data.}
\usage{
ir2r.sp(object, row.max = NULL, col.max = NULL)
}
\arguments{
\item{object}{Irregular spatial data with Row and Col.}

\item{row.max}{Total row numbers for regular data.}

\item{col.max}{Total column numbers for regular data.}
}
\value{
\code{object}	Irregular spatial data, Must have 'Row' and 'Col' as the first two column.
}
\description{
\code{ir2r.sp} This function changes irregular spatial data to regular data for 
further plot with spd.plot().
}
\examples{
\dontrun{
library(AAFMM)

data(ir.sp)

ir.sp2<-ir.sp[,5:16] # order: Row,Col,h05,cw05,...
#ir.sp2<-subset(ir.sp,select=c(Row,Col,h05,cw05))

sp1<-ir2r.sp(ir.sp2,row.max=10)
sp2<-ir2r.sp(ir.sp2,col.max=20)
sp3<-ir2r.sp(ir.sp2,row.max=10,col.max=20)

}
}
\references{
Yuanzhen Lin. R & ASReml-R Statistics. China Forestry Publishing House. 2016 
AAFMM website:https://github.com/yzhlinscau/AAFMM
}
\author{
Yuanzhen Lin <yzhlinscau@163.com>
}
