% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/met.biplot.R
\name{met.biplot}
\alias{met.biplot}
\title{Biplot asreml-MET results.}
\usage{
met.biplot(object, siteN, VarietyN, faN,dSco.u=NULL,dLam.u=NULL,asrV=3)
}
\arguments{
\item{object}{asreml factor analytic results for MET.}

\item{siteN}{Total trial site numbers.}

\item{VarietyN}{Total variety numbers.}

\item{faN}{Total factor numbers.}

\item{dSco.u}{Least score of Variety breeding value.}

\item{dLam.u}{Least distance from center.}

\item{asrV}{asreml version 3(default V3), or 4 (new version V4).}
}
\description{
\code{met.biplot} This function biplots MET factor analytic results from asreml 
to find the relation of trial sites and the best variety suitable to trial sites.
}
\examples{
\dontrun{

library(AAFMM)
data(MET)
MET$yield<-0.01*MET$yield

## for ASReml-R V3.0
library(asreml)
met.asr<-asreml(yield~Loc, random=~ Genotype:fa(Loc,2), 
                rcov=~ at(Loc):ar1(Col):ar1(Row), 
                data=MET, maxiter=50)

met.biplot(met.asr,6,36,2)
met.biplot(met.asr,siteN=6,VarietyN=36,faN=2,
                   dSco.u=1.8,dLam.u=1.0)


## for ASReml-R V4.0
library(asreml4)
met.asr<-asreml(yield~Loc, random=~ Genotype:fa(Loc,2), 
                residual=~ dsum(~ar1(Col):ar1(Row)|Loc), 
                data=MET, maxiter=50)

met.biplot(met.asr,6,36,2,asrV=4)
met.biplot(met.asr,siteN=6,VarietyN=36,faN=2,
             dSco.u=1.8,dLam.u=1.0,asrV=4)

}


}
\references{
Yuanzhen Lin. R & ASReml-R Statistics. China Forestry Publishing House. 2016 
AAFMM website:https://github.com/yzhlinscau/AAFMM
}
\author{
Yuanzhen Lin <yzhlinscau@163.com>
}
