% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/met.corr.R
\name{met.corr}
\alias{met.corr}
\title{asreml-MET corr matrix.}
\usage{
met.corr(object, site,faN=1, kn = NULL)
}
\arguments{
\item{object}{asreml factor analytic results for MET.}

\item{site}{Variable site of data.}

\item{faN}{Factor numbers,1(default).}

\item{kn}{Site cluster group numbers.}
}
\value{
\code{object}	asreml factor analytic results for MET, met.asr.
\code{site}	variable site of data, MET$Site.
\code{faN}	factor numbers,1(default).
\code{kn}	cluster group numbers,3.
}
\description{
\code{met.corr} This function calculate var/cov/corr from asreml MET factor analytic 
results to further research the relation of trial sites.
}
\examples{
\dontrun{
library(AAFMM)
data(MET)
MET$yield<-0.01*MET$yield

## for ASReml-R V3.0 
library(asreml)

met.asr1<-asreml(yield~Loc, random=~ Genotype:fa(Loc,2), 
                rcov=~ at(Loc):ar1(Col):ar1(Row), 
                data=MET, maxiter=50)

met.corr(met.asr1,MET$Loc,faN=2)

 
## for ASReml-R V4.0 
library(asreml4)

met.asr2<-asreml(yield~Loc, random=~ Genotype:fa(Loc,2), 
                residual=~ dsum(~ar1(Col):ar1(Row)|Loc), 
                data=MET, maxiter=50)

met.corr(met.asr2,MET$Loc,faN=2)
}

}
\references{
Yuanzhen Lin. R & ASReml-R Statistics. China Forestry Publishing House. 2016 
AAFMM website:https://github.com/yzhlinscau/AAFMM
}
\author{
Yuanzhen Lin <yzhlinscau@163.com>
}
