% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/met.plot.R
\name{met.plot}
\alias{met.plot}
\title{Plot asreml-MET data.}
\usage{
met.plot(object, plot.title = NULL)
}
\arguments{
\item{object}{MET data.}

\item{plot.title}{MET plot title.}
}
\description{
\code{met.plot} This function plots MET data for further 
factor analytic by asreml to find the relation of trial sites, etc.
}
\examples{
\dontrun{
library(AAFMM)

data(MET)

# example 1
# variable order: genotype,yield,site,row,col
MET2<-MET[,c(1,9,2,4:5)]
 
met.plot(MET2)

# example 2
# variable order on MET2: Rep, Block
MET3<-MET[,c(1,9,2,4:7)] 

met.plot(MET3,"My met trials")
}

}
\references{
Yuanzhen Lin. R & ASReml-R Statistics. China Forestry Publishing House. 2016 
AAFMM website:https://github.com/yzhlinscau/AAFMM
}
\author{
Yuanzhen Lin <yzhlinscau@163.com>
}
