% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.comp.R
\name{model.comp}
\alias{model.comp}
\title{Model comparison for asreml.}
\usage{
model.comp(m1=NULL,m2=NULL,Nml=NULL,mulM=NULL,LRT=NULL,rdDF=NULL)
}
\arguments{
\item{m1}{Model 1 asreml results,such as "m1.asr".}

\item{m2}{Model 2 asreml results.}

\item{Nml}{A vector with more than 2 asreml results, such as "c(m1,m2,m3,m4)".}

\item{mulM}{mulM is "TRUE" for a vector's comparison with multi-models, default (FALSE).}

\item{LRT}{Value TRUE for Likelihood ratio test (LRT), default (FALSE) for no LRT.}

\item{rdDF}{Value TRUE to control Ddf minus 0.5, default FALSE for LRT with Ddf. 
If TRUE, Ddf would minus 0.5, ie, Ddf=Ddf-0.5.}

\item{asrV}{asreml version 3(default, V3), or 4 (new version V4).}
}
\description{
\code{model.comp} This function would compare models with different random structure 
under the same fixed factors.
}
\examples{
\dontrun{

library(AAFMM)
df<-PrSpa

## for ASReml-R V3.0
library(asreml)
fm<-asreml(h5~ 1+Rep,random=~ Fam, 
           subset=Spacing=='3',data=df,maxit=40)

fm1a<-asreml(cbind(dj,h5)~ trait+trait:Rep, 
             random=~ us(trait):Fam, 
             rcov=~units:us(trait),
             subset=Spacing=='3',data=df,maxit=40)
             
fm1b<-asreml(cbind(dj,h5)~ trait+trait:Rep, 
             random=~ diag(trait):Fam, 
             rcov=~units:us(trait),
             subset=Spacing=='3',data=df,maxit=40)

fm3a<-asreml(cbind(dj,h3,h5)~ trait+trait:Rep, 
             random=~ diag(trait):Fam, 
             rcov=~units:us(trait),
             subset=Spacing=='3',data=df,maxit=40)

fm3b<-asreml(cbind(dj,h3,h5)~ trait+trait:Rep, 
             random=~ diag(trait):Fam, 
             rcov=~units:us(trait),
             subset=Spacing=='3',data=df,maxit=40)

#####   model comparison    #####
model.comp(m1=fm1a,m2=fm1b)
model.comp(m1=fm1a,m2=fm1b,LRT=TRUE)
model.comp(m1=fm1a,m2=fm1b,LRT=TRUE,rdDF=TRUE)

model.comp(Nml=c(fm3a,fm3b,fm1a,fm1b,fm),mulM=TRUE)
model.comp(Nml=c(fm3a,fm3b,fm1a,fm1b,fm),mulM=TRUE,LRT=TRUE)


## for ASReml-R V4.0
library(asreml4)

fm<-asreml(h5~ 1+Rep,random=~ Fam, 
           subset=Spacing=='3',data=df,maxit=40)

fm1a<-asreml(cbind(dj,h5)~ trait+trait:Rep, 
             random=~ us(trait):Fam, 
             residual=~units:us(trait),
             subset=Spacing=='3',data=df,maxit=40)
             
fm1b<-asreml(cbind(dj,h5)~ trait+trait:Rep, 
             random=~ diag(trait):Fam, 
             residual=~units:us(trait),
             subset=Spacing=='3',data=df,maxit=40)

fm3a<-asreml(cbind(dj,h3,h5)~ trait+trait:Rep, 
             random=~ diag(trait):Fam, 
             residual=~units:us(trait),
             subset=Spacing=='3',data=df,maxit=40)

fm3b<-asreml(cbind(dj,h3,h5)~ trait+trait:Rep, 
             random=~ diag(trait):Fam, 
             residual=~units:us(trait),
             subset=Spacing=='3',data=df,maxit=40)

#####   model comparison    #####
model.comp(m1=fm1a,m2=fm1b,asrV=4)
model.comp(m1=fm1a,m2=fm1b,LRT=TRUE,asrV=4)
model.comp(m1=fm1a,m2=fm1b,LRT=TRUE,rdDF=TRUE,asrV=4)

model.comp(Nml=c(fm3a,fm3b,fm1a,fm1b,fm),mulM=TRUE,asrV=4)
model.comp(Nml=c(fm3a,fm3b,fm1a,fm1b,fm),mulM=TRUE,LRT=TRUE,asrV=4)
}

}
\references{
Yuanzhen Lin. R & ASReml-R Statistics. China Forestry Publishing House. 2016 
AAFMM website:https://github.com/yzhlinscau/AAFMM
}
\author{
Yuanzhen Lin <yzhlinscau@163.com>
}
