% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pin.R
\name{pin}
\alias{pin}
\alias{pin.asreml}
\alias{pin.remlf90}
\title{Count error for h2 and corr.}
\usage{
pin(object, formula, asrV, signif, corN, Rdf, digit, vres)

\method{pin}{asreml}(object, formula = NULL, asrV = 3, signif = FALSE,
  corN = NULL, Rdf = FALSE, digit = 3, vres = FALSE)

\method{pin}{remlf90}(object, formula = NULL, signif = FALSE, digit = 3,
  vres = FALSE)
}
\arguments{
\item{object}{asreml or breedR results.}

\item{formula}{formula for h2 or corr.}

\item{asrV}{Index for asreml version, 3(default),or 4(new version).}

\item{signif}{Index to output signif levels, F(default) for non-signif.}

\item{corN}{Number of corr, 1(default).}

\item{Rdf}{Index to output results to vector, F(default) for non-vector output.}

\item{digit}{Index for decimal number, 3(default).}

\item{vres}{Index(T) to return results in vectors, F(default) for direct results.}
}
\value{
the result is returned directly.
}
\description{
\code{pin} This function counts standard error(se) for heritability(h2) and 
corr value and also outputs significent level for corr value in asreml and breedR package.
}
\details{
Count error for h2 and corr value, also outputs significent level.
}
\examples{
\dontrun{
## working for breedR package
library(breedR)
library(AAFMM)

data(globulus)
res.animal <- remlf90(fixed = phe_X ~ 1,
                      random = ~ gg,
                      genetic = list(model = 'add_animal',
                      pedigree = globulus[, 1:3],
                      id = 'self'),
                      data = globulus)
                      
pin(res.animal,h2~V2/(V1+V2+V3))
} 

\dontrun{
## working for asreml or asreml4 package
library(AAFMM)
data(PrSpa)
df<-PrSpa

## when works for ASReml-R V3.0
library(asreml) # V3.0

# exmaple 1.1 single trait model
fm1<-asreml(h5~1+Rep, random=~Fam, 
             subset=Spacing=='3',data=df)

summary(fm)$varcomp[,1:3]

pin(fm1, h2 ~4*V1/(V1+V2))
pin(fm1, h2 ~4*V1/(V1+V2),Rdf=TRUE)

# exmaple 1.2 us model
fm2<-asreml(cbind(dj,h5)~ trait+trait:Rep, 
            random=~ us(trait):Fam, 
            rcov=~units:us(trait),
            subset=Spacing=='3',data=df,maxit=40)
                        
summary(fm2)$varcomp[,1:3]

pin(fm2, h2_A ~ 4 * V1/(V1+V5)) # heritability for trait A
pin(fm2, h2_B ~ 4 * V3/(V3+V7)) # heritability for trait B

# genetic corr
pin(fm2, gCORR ~ V2/sqrt(V1*V3),signif=TRUE) 

# phenotype corr
pin(fm2, pCORR ~ (V2+V6)/sqrt((V1+V5)*(V3+V7)),signif=TRUE) 

# exmaple 1.3 corr model
fm3<-asreml(cbind(dj,h3,h5)~ trait+trait:Rep, 
            random=~ corgh(trait):Fam, 
            rcov=~units:us(trait),
            subset=Spacing=='3',data=df,maxit=40)
                        
summary(fm3)$varcomp[,1:3]
pin(fm3,corN=3)

## when works for ASReml-R V4.0
library(asreml4)

# exmaple 2.1 single trait model, default gamma parameterization.
fm1b<-asreml(h5~1+Rep, random=~Fam, subset=Spacing=='3',data=df)

summary(fm)$varcomp[,1:3]

pin(fm1b, h2 ~4*V1/(V1+V2),asrV=4)
pin4(fm1b, h2 ~4*V1/(V1+V2)) # same results, but with pin4()

# The same model with fm1b, but with sigma parameterization.
fm1c<-asreml(h5~1+Rep, random=~Fam, 
               residual=~idv(units),
               subset=Spacing=='3',data=df)
               
summary(fm1c)$varcomp[,1:3]

pin(fm1c, h2 ~4*V1/(V1+V3),asrV=4)
pin4(fm1c, h2 ~4*V1/(V1+V3))

# exmaple 2.2 us model
fm2b<-asreml(cbind(h3,h5)~ trait+trait:Rep, 
            random=~ us(trait):Fam, 
            residual=~units:us(trait),
            subset=Spacing=='3',data=df,maxit=40)
            
summary(fm2b)$varcomp[,1:3]

# heritability for trait A
pin(fm2b, h2_A ~ 4 * V1/(V1+V5),asrV=4) 
pin4(fm2b, h2_A ~ 4 * V1/(V1+V5)) 

# heritability for trait B
pin(fm2b, h2_B ~ 4 * V3/(V3+V7),asrV=4) 
pin4(fm2b, h2_B ~ 4 * V3/(V3+V7))

# genetic corr
pin(fm2b, gCORR ~ V2/sqrt(V1*V3),signif=TRUE,asrV=4)
pin4(fm2b, gCORR ~ V2/sqrt(V1*V3),signif=TRUE)

# phenotype corr
pin(fm2b, pCORR ~ (V2+V6)/sqrt((V1+V5)*(V3+V7)),signif=TRUE,asrV=4) 
pin4(fm2b, pCORR ~ (V2+V6)/sqrt((V1+V5)*(V3+V7)),signif=TRUE) 

# exmaple 2.3 corr model
fm3b<-asreml(cbind(h3,h4,h5)~ trait+trait:Rep, 
            random=~ corgh(trait):Fam, 
            residual=~units:us(trait),
            subset=Spacing=='3',data=df,maxit=40)
                        
summary(fm3b)$varcomp[,1:3]

pin(fm3b,corN=3,asrV=4) 
pin4(fm3b,corN=3) 
}


}
\references{
Yuanzhen Lin. R & ASReml-R Statistics. China Forestry Publishing House. 2016 
AAFMM website:https://github.com/yzhlinscau/AAFMM
}
\author{
Yuanzhen Lin <yzhlinscau@163.com>
}
