% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot1.R
\name{plot1}
\alias{plot1}
\alias{plot1.remlf90}
\alias{plot1.group}
\title{Plotting test trait's norm or multi-comparison}
\usage{
plot1(object, mulT, x.lbls, y.lbls, y.zero)

\method{plot1}{remlf90}(object, mulT = FALSE)

\method{plot1}{group}(object, x.lbls = NULL, y.lbls = NULL, y.zero = NULL)
}
\arguments{
\item{object}{an object of breedR result or multi-comparison.}

\item{mulT}{multi-trait model(default, FALSE).}
}
\value{
the result is returned directly.
}
\description{
\code{plot1} This function plots test trait's norm for 
    breedR object or multi-comparisons for agricolae.
}
\details{
Test trait's norm for breedR object,similar to asreml.
}
\examples{
## 1 working for agricolae package
library(AAFMM)
library(agricolae)

data(sweetpotato)
model<-aov(yield~virus,data=sweetpotato)
comparison<- LSD.test(model,"virus",alpha=0.01)

plot1(comparison,x.lbls="virus")

\dontrun{
## 2 working for breedR package
library(breedR)
library(AAFMM)

res.animal <- remlf90(fixed = phe_X ~ 1,
                      random = ~ gg,
                      genetic = list(model = 'add_animal',
                      pedigree = globulus[, 1:3],
                      id = 'self'),
                      data = globulus)
                      
plot1(res.animal)
}

}
\references{
AAFMM website:https://github.com/yzhlinscau/AAFMM
}
\author{
Yuanzhen Lin <yzhlinscau@163.com>
}
