% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.file.R
\name{read.example}
\alias{read.example}
\title{read file list}
\usage{
read.example(package, setpath = FALSE)
}
\arguments{
\item{package}{package name.}

\item{setpath}{Whether set working directory under package, FALSE(default).}
}
\value{
this returned a path or file list.
}
\description{
\code{read.example} This function read file list under one package, or
sets working directory under package.
}
\examples{
library(AAFMM)
# read file list under a package
AAFMM::read.example(package = "AAFMM")

# set working directory under a package
AAFMM::read.example(package = "AAFMM", setpath = TRUE)
getwd()

}
\references{
Yuanzhen Lin. R & ASReml-R Statistics. China Forestry Publishing House. 2016
AAFMM website:https://github.com/yzhlinscau/AAFMM
}
\seealso{
See Also as \code{\link{read.file}}, \code{\link{fdata}}
}
\author{
Yuanzhen Lin <yzhlinscau@163.com>
}
