% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp.batch.R
\name{snp.batch}
\alias{snp.batch}
\title{SNP batch analysis in asreml.}
\usage{
snp.batch(data,SNP.N,traitN,FMod=NULL,maxit=NULL,SNP.detail=FALSE,SNP.figure=FALSE,SNP.signif=TRUE,alpha=0.05)
}
\arguments{
\item{data}{asreml dataset.}

\item{SNP.N}{A vector with sites of all SNPs.}

\item{traitN}{A vector with sites of all traits.}

\item{FMod}{Fixed mode, y~x(default).}

\item{maxit}{Maximum number of iterations, 20(default).}

\item{SNP.detail}{Index to output information of every SNP type, F(default) for non-SNP-detail output.}

\item{SNP.figure}{Index to output boxplot of every SNP type, F(default) for non-figure output.}

\item{SNP.signif}{Index to output signif results of every SNP type, TRUE(default) for only significant output.}

\item{alpha}{Signif levels of every SNP type, 0.05(default).}
}
\description{
\code{snp.batch} This function carries out batch analysis for SNP effects with trait in asreml package.
}
\examples{
\dontrun{
library(AAFMM)
data(SNP)
df<-SNP #str(df)

library(asreml) # or library(asreml4)

snp.batch(df,SNP.N=2:4,traitN=c(5:6))
snp.batch(df,SNP.N=2:4,traitN=c(5:6), SNP.detail=TRUE)
snp.batch(df,SNP.N=2:4,traitN=c(5:6), SNP.figure=TRUE)
}

}
\references{
Yuanzhen Lin. R & ASReml-R Statistics. China Forestry Publishing House. 2016 
AAFMM website:https://github.com/yzhlinscau/AAFMM
}
\author{
Yuanzhen Lin <yzhlinscau@163.com>
}
