% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp.R
\docType{data}
\name{sp}
\alias{sp}
\title{sp}
\format{one dataset with 10 variables
\describe{
 \item{Tree}{Tree ID}
 \item{Mum}{Mum ID, 34 levels}
 \item{Dad}{Dad ID, 32 levels}
 \item{Fam}{Family ID, 42 levels}
 \item{Row}{Row number}
 \item{Col}{Column number}
 \item{Rep}{Replication level with 7}
 \item{Plot}{Plot level with 5}
 \item{dbh10}{a numeric vector}
 \item{t2}{a numeric vector}
}}
\usage{
data(sp)
}
\description{
An example dataset of spatial analysis.
}
\keyword{datasets}
